Rods partial GUI and functionality + refactorization
GUI:

- added a new screen for iRODS. This screen can be used for entering iRODS
credentials;
- added two radio buttons for selecting between the screen for iRODS or for
Dbx.

Functionality:
- connection to iRODS server;
- possibility to browse files from local station and from remote rods server.

Refactorization:
- added an interface for cloud operations. Both iRODS  and Dbx clients have
to implement this interface. This allows the usage of the generic interface
in the MyCloudJ_ class, while the backend is handled in specific classes
for Dbx and iRODS;
- actions listeners for buttons are implemented in internal classes from
MyCloudJ_ class. This is in contrast with defining action listerners in
anonymous classes;
- created a new method for drawing the GUI and added the component drawing
inside this method. This method is called from the main run method of the
MyCloudJ_ class;
- javadoc for MyCloudJ fields. TODO: javadoc for methods and methods fields.
- added the package rodsUtils which handles the backend for the connection
with the rods server;
- added the package generalUtils which handles commun backend operations:
e.g.: determine O.S type;

P.S: I know that I should have split this commit in more little commits :)