File browsing capability for iRODS
Added the functionality for file browsing on a remote iRODS server.

Also, changed how a node's name from the JTree is displayed. Instead
of displaying the full path of a node like the path to the root directory
node we now display only the last component of the path.
Ex: /test/node_name => node_name
The reason for this choice is the hierarchically structure of the JTree.