Refactor: Remove GUI processing from cloud specific classes
This commit removes the processing of the browsing trees from the cloud
backend specific classes. Before the refactoring, the cloud classes were
iterating through the files of a directory and were filing the JTree. Now,
we just build a list with the child for a folder and the filing of the JTree
is done in the GUI specific class.