Modify behaviour for updating browsing trees after upload
The old behaviour for updating the browsing trees after a successful
upload is described in 09681e14d73ca1c66ca56238af4241ffa2699cc1.

While the old behaviour was to update the browsing trees based on the
currently _selected_ node the new behaviour is to update the browsing
trees based on the _folder where we've done the upload_.

The download tree will be expanded only if:
- the folder where we've done the upload is expanded or:
- the folder where we've done the upload contains only the uploaded
file and the parent of the folder where we've done the upload is
expanded.

The upload tree will be expanded only if:
- the folder/parent of the folder where we've done the upload is
expanded.