Redesign of file transfer backend
This commit keeps the consumer-producer pattern but it implements it
using the an Executor and the FutureTask class. The FutureClass wraps
a Callable object which can be interrupted using the cancel method. We
cancel a file transfer if the user disconnects from iRODS in the middle
of a file transfer. Later, I plan to use this class for pausing a file
transfer.

By default, Jargon API doesn't contain a method for canceling the file
transfer so I needed to wrap the file transfer operations in a FutureTask
object. More details: https://github.com/DICE-UNC/jargon/issues/90