iRODS: Add delete file functionaity
A right-click popup menu is available in the cloud file browsing tree.
User has the possibility to delete a file from iRODS.

Unfortunately, Dbx Java API doesn't have yet support for delete operation.
User is informed about the lack of capabilities when he tries to execute
an operation which is not supported.