Fix:  Correctly update the download browsing tree
Before this commit, when the user have successfully uploaded a file, the
incorrect node from the download browsing tree was updated with the newly
uploaded file.

Steps for reproducing the behaviour:
P1: select the radio button for download
P2: open the download browsing tree and expand a folder
P3: select the radio button for upload
P4: upload a file in a different folder than the one selected in the
download browsing tree
P5: select the radio button for download
P6: check that the node updated in the download browsing tree is the
one select in P3.

Fixes: a7460d15e8a8c60e06811cdf52f08db48bccfe0e