Open downloaded files using double click events
User can open downloaded files by double click on the progress table.
The default opener might not be able to open the file because the file is
not in a compatibile format.

Unfortunately, in Java 7 there is a bug when trying to determine file type:
http://bugs.java.com/bugdatabase/view_bug.do?bug_id=7096063
so instead of doing that buggy verification we'll try to open any file.